loginpageUI

+++ src/pages/SignUpPage/SignUpPage.tsx
@@ -39,6 +39,7 @@ const SignUpDetails = (state, action) => {
 function SignUpPage({ props: setlogin }: Props) {
 	const [passwordCheckError, setPasswordCheckError] = useState(false);
 	const [confirmPassword, setConfirmPassword] = useState("");
+	const [comparePassword, setComparePassword] = useState(false);
 	const [state, dispatch] = useReducer(SignUpDetails, {
 		email: "",
 		password: "",
@@ -94,38 +95,44 @@ function SignUpPage({ props: setlogin }: Props) {
 		setPasswordType((prev) => (prev === "password" ? "text" : "password"));
 	};
 
-	const onSubmittFunc = async () => {
+	const onSubmittFunc = async (e) => {
+		e.preventDefault();
 		try {
-			var object = {
-				email: state.email,
-				password: confirmPassword,
-				firstName: state.firstName,
-				lastName: state.lastName,
-			};
-			var res = await fetch({
-				method: "post",
-				url: "/api/auth/signup",
-				data: object,
-			});
-			console.log(res);
-			var token = res?.encodedToken;
-			localStorage.setItem(VAR_ENCODE_TOKEN, token);
-			var user = res?.createdUser;
-			var userId = res.createdUser._id;
-			localStorage.setItem(VAR_USER_ID, userId);
-			auth.loginUser(
-				{
-					email: res.createdUser.email,
-					firstName: res.createdUser.firstName,
-					lastName: res.createdUser.lastName,
-				},
-				() => {
-					navigate(from?.pathname || "/home", { replace: true });
-				}
-			);
-			setlogin(true);
-			setmodalToggle(false);
-			Toast("Signed In!")
+			if (state.email.trim() === "" || confirmPassword.trim() === "" || state.firstName.trim() === "" || state.lastName.trim()=== "") {
+				// Alert("error", "Input cannot be blank");
+				Toast("Input cannot be blank")
+			} else {
+				var object = {
+					email: state.email,
+					password: confirmPassword,
+					firstName: state.firstName,
+					lastName: state.lastName,
+				};
+				var res = await fetch({
+					method: "post",
+					url: "/api/auth/signup",
+					data: object,
+				});
+				console.log(res);
+				var token = res?.encodedToken;
+				localStorage.setItem(VAR_ENCODE_TOKEN, token);
+				var user = res?.createdUser;
+				var userId = res.createdUser._id;
+				localStorage.setItem(VAR_USER_ID, userId);
+				auth.loginUser(
+					{
+						email: res.createdUser.email,
+						firstName: res.createdUser.firstName,
+						lastName: res.createdUser.lastName,
+					},
+					() => {
+						navigate(from?.pathname || "/home", { replace: true });
+					}
+				);
+				setlogin(true);
+				setmodalToggle(false);
+				Toast("Signed In!")
+			}
 		} catch (error: any) {
 			console.log(error.message);
 			Toast(error.message)
@@ -144,26 +151,66 @@ function SignUpPage({ props: setlogin }: Props) {
 			<div className="signup-body-container"
 				onClick={(event: React.MouseEvent<HTMLElement>) => StopPropogation(event)} >
 
-				{/* <img src={Signup} className="signupImage" alt="signupImg" /> */}
-				<div className="signup-container">
-					<div className="title-header">
-						<p style={{fontSize: "10px"}} >
-							Create your profile and get access
-							to create notes and maintain them
-						</p>
-					</div>
-					<div className="signup-credential-container">
-						<input
-							type="email"
-							placeholder="Email Address - xyz@gmail.com"
-							onChange={(e) => dispatch({ email: e.target.value })}
-						/>
-					</div>
-					<div className="signup-credential-container">
-						<div className="password-holder">
+					<form onSubmit={(e) => onSubmittFunc(e)} className="signup-container">
+						<div className="title-header">
+							<p className="lg-txt" >
+								Create your profile and get access
+								to create notes and maintain them
+							</p>
+						</div>
+						<div className="signup-credential-container">
 							<input
-								type={passwordType}
-								placeholder="Password"
+								type="email"
+								placeholder="Email Address - xyz@gmail.com"
+								onChange={(e) => dispatch({ email: e.target.value })}
+							/>
+						</div>
+						<div className="signup-credential-container">
+							<div className="password-holder">
+								<input
+									type={passwordType}
+									placeholder="Password"
+									name=""
+									id=""
+									style={{
+										borderColor: passwordCheckError ? "red" : "black",
+										outlineColor: passwordCheckError ? "red" : "black",
+									}}
+									onChange={(e) => {
+										PasswordCheck(e.target.value);
+									}}
+								/>
+								{passwordType === "password" ? (
+									<span
+										className="material-icons-round"
+										onClick={() => PasswordVisibilityHandler()}
+									>
+										visibility
+									</span>
+								) : (
+									<span
+										className="material-icons-round"
+										onClick={() => PasswordVisibilityHandler()}
+									>
+										visibility_off
+									</span>
+								)}
+							</div>
+							<p className="error">
+								{confirmPassword.length > 0 && confirmPassword.length < 7
+									? "password should be minimum 7 letter"
+									: ""}
+							</p>
+							<p className="error">
+								{(confirmPassword.length > 0 &&
+									passwordCheckError ) ||
+									"Password should contain a number, alphabet & special character"}
+							</p>
+						</div>
+						<div className="signup-credential-container">
+							<input
+								type="password"
+								placeholder="Confirm Password"
 								name=""
 								id=""
 								style={{
@@ -171,87 +218,46 @@ function SignUpPage({ props: setlogin }: Props) {
 									outlineColor: passwordCheckError ? "red" : "black",
 								}}
 								onChange={(e) => {
-									PasswordCheck(e.target.value);
+									setComparePassword(e.target.value !== confirmPassword);
 								}}
+								disabled={
+									passwordCheckError && confirmPassword.length >= 7 ? false : true
+								}
 							/>
-							{passwordType === "password" ? (
-								<span
-									className="material-icons-round"
-									onClick={() => PasswordVisibilityHandler()}
-								>
-									visibility
-								</span>
-							) : (
-								<span
-									className="material-icons-round"
-									onClick={() => PasswordVisibilityHandler()}
-								>
-									visibility_off
-								</span>
-							)}
+							<p className="error">
+								{comparePassword && "Confirm Password Should match Password"}
+							</p>
 						</div>
-						<p className="error">
-							{confirmPassword.length > 0 && confirmPassword.length < 7
-								? "password should be minimum 7 letter"
-								: ""}
-						</p>
-						<p className="error">
-							{confirmPassword.length > 0 &&
-								!passwordCheckError &&
-								"Password should contain a number, alphabet & special character"}
-						</p>
-					</div>
-					<div className="signup-credential-container">
-						<input
-							type="password"
-							placeholder="Confirm Password"
-							name=""
-							id=""
-							style={{
-								borderColor: passwordCheckError ? "red" : "black",
-								outlineColor: passwordCheckError ? "red" : "black",
-							}}
-							onChange={(e) => {
-								setPasswordCheckError(e.target.value !== confirmPassword);
-							}}
-							disabled={
-								passwordCheckError && confirmPassword.length >= 7 ? false : true
-							}
-						/>
-						<p className="error">
-							{passwordCheckError && "Confirm Password Should match Password"}
-						</p>
-					</div>
-					<div className="signup-credential-container">
-						<input
-							type="email"
-							placeholder="First Name"
-							onChange={(e) => dispatch({ firstName: e.target.value })}
-						/>
-						{/* </div>
-        <div className="signup-credential-container"> */}
-						<input
-							type="email"
-							placeholder="Last Name"
-							onChange={(e) => dispatch({ lastName: e.target.value })}
-						/>
-					</div>
-					<div className="signup-remember-container">
-						
-							<input type="checkbox" name="" id="" />I accept all Terms &
-							Conditions
-						
-					</div>
-					<div className="signup-btn-container">
-						<button className="btn signup-action-btn" onClick={onSubmittFunc}>
-							Signup
-						</button>
-					</div>
-					<span className="signup-footer" onClick={() => setlogin(true)}>
+						<div className="signup-credential-container">
+							<input
+								type="email"
+								placeholder="First Name"
+								onChange={(e) => dispatch({ firstName: e.target.value })}
+							/>
+							{/* </div>
+			<div className="signup-credential-container"> */}
+							<input
+								type="email"
+								placeholder="Last Name"
+								onChange={(e) => dispatch({ lastName: e.target.value })}
+							/>
+						</div>
+						<div className="signup-remember-container">
+							
+								<input type="checkbox" name="" id="" />I accept all Terms &
+								Conditions
+							
+						</div>
+						<div className="signup-btn-container">
+							<button className="btn signup-action-btn" type="submit" >
+								Signup
+							</button>
+						</div>
+						<span className="signup-footer cursive" onClick={() => setlogin(true)}>
 						Already have an Account{" "}
 						<span className="material-icons-round">navigate_next</span>
-					</span>
-				</div>
+						</span>
+					</form>
 			</div>
 		</FullPageModal>
 	);
